Rebuilds Timer component
- Uses internal state over redux. At this time I'm a little
inexperienced with react hooks and the errors produced
from using redux was not worth the headache of pushing
this components state to a global state store. It feels better using
independent internal state. Keeping redux for now because I like
the tool and I'd like to at least play with in the react space, but it
doesn't feel like the right tool for this job going forward.

- Uses react component class over function