Fixes interval bug
First time using setInterval in React and ran into a problem. When
the component is rendered all the values of the interval are only
evaluated once so regardless of state change the state is always
fixed to that moment of render.

This change uses the React `useEffect` hook to bind the effect
to a few state properties to be re-evaluated when those change.